pyproject.toml
[project]
name = "anomaly-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "torch>=2.9.1",
]
